Move ring-backed signature keys to crypto::ring
