+++ rustls/src/crypto/ring/mod.rs
@@ -20,6 +20,9 @@ pub(crate) mod quic;
 pub(crate) mod tls12;
 pub(crate) mod tls13;
 
+/// Using software keys for authentication.
+pub mod sign;
+
 /// Default crypto provider.
 #[derive(Debug)]
 pub struct Ring;
