+++ rustls/src/lib.rs
@@ -491,10 +491,10 @@ pub use crypto::ring::kx_group;
 
 /// Message signing interfaces and implementations.
 pub mod sign {
-    pub use super::crypto::signer::{
-        any_ecdsa_type, any_eddsa_type, any_supported_type, CertifiedKey, RsaSigningKey, Signer,
-        SigningKey,
+    pub use crate::crypto::ring::sign::{
+        any_ecdsa_type, any_eddsa_type, any_supported_type, RsaSigningKey,
     };
+    pub use crate::crypto::signer::{CertifiedKey, Signer, SigningKey};
 }
 
 #[cfg(feature = "quic")]
