+++ rustls/src/server/handy.rs
@@ -1,3 +1,4 @@
+use crate::crypto::ring;
 use crate::dns_name::DnsNameRef;
 use crate::error::Error;
 use crate::limited_cache;
@@ -103,7 +104,7 @@ impl AlwaysResolvesChain {
         chain: Vec<CertificateDer<'static>>,
         priv_key: &PrivateKeyDer<'_>,
     ) -> Result<Self, Error> {
-        let key = sign::any_supported_type(priv_key)
+        let key = ring::sign::any_supported_type(priv_key)
             .map_err(|_| Error::General("invalid private key".into()))?;
         Ok(Self(Arc::new(sign::CertifiedKey::new(chain, key))))
     }
